+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Entering the nether now takes 2.5s instead of 4s (configurable)
+* Added config option to prevent equipment from dropping enchanted
 * Lowered hardness of blocks below sea level (~~+64%~~ -> +50% at Y=0 and below)
 
 ## 3.14.8